requirements.txt
azure-cosmos
azure-storage-blob
pymongo
python-dotenv